Initial group project commit with minimal changes.
This initial version is supposed to mimic the final product of the group
project in CSC 439 at NKU. As per project specifications, this is NOT
supposed to be a FULL Blackjack game, but it should be close. Therefore,
while there are JUnit tests, there may be many bugs and strange
behavior.